Split HIR definition collection (#1300)
Move type-name collection, aggregate definitions, trait type-use validation, and recursive aggregate checks from hir.rs into hir/definitions.rs while preserving compiler.hir behavior. Update the compiler source monolith reporter and ratchet ceilings for the reduced HIR facade line count.

Refs #1254.
